+++ rasa/core/training/interactive.py
@@ -655,22 +655,34 @@ async def _request_action_from_user(
 def _request_export_info() -> Tuple[Text, Text, Text]:
     """Request file path and export stories & nlu data to that path"""
 
+    def validate_story_export_file(path: Text) -> bool:
+        return path and path.endswith(".md")
+
+    def validate_nlu_export_file(path: Text) -> bool:
+        return path and (path.endswith(".md") or path.endswith(".json"))
+
+    def validate_domain_export_file(path: Text) -> bool:
+        return path and (path.endswith(".yaml") or path.endswith(".yml"))
+
     # export training data and quit
     questions = questionary.form(
         export_stories=questionary.text(
             message="Export stories to (if file exists, this "
             "will append the stories)",
             default=PATHS["stories"],
+            validate=validate_story_export_file
         ),
         export_nlu=questionary.text(
             message="Export NLU data to (if file exists, this will "
             "merge learned data with previous training examples)",
             default=PATHS["nlu"],
+            validate=validate_nlu_export_file
         ),
         export_domain=questionary.text(
             message="Export domain file to (if file exists, this "
             "will be overwritten)",
             default=PATHS["domain"],
+            validate=validate_domain_export_file
         ),
     )
 
@@ -778,8 +790,7 @@ async def _write_nlu_to_file(
     try:
         previous_examples = load_data(export_nlu_path)
     except Exception as e:
-        logger.exception("An exception occurred while trying to load the NLU data.")
-
+        logger.error("An exception occurred while trying to load the NLU data.")
         export_nlu_path = questionary.text(
             message="Could not load existing NLU data, please "
             "specify where to store NLU data learned in "