+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Removed
 -------
 - revert the stripping of trailing slashes in endpoint URLs since this can lead to
   problems in case the trailing slash is actually wanted
+- validate export paths in interactive learning
 
 Fixed
 -----